Minor edits to makefile and .bib
Makefile: got rid of a couple of cases of lines starting with non-tab.
Harmless, but my editor moans about them.

Also, I changed /Documents/ to /documents/ in ivoatex/ivoabib.bib, since the
latter appears now to be canonical (in the sense that ivoa.net/Documents is now
rewritten by that server), and since the first two references in the document
had different capitalisations.


git-svn-id: https://volute.g-vo.org/svn/trunk/projects/ivoapub/ivoatex@2819 4e403e14-6840-0410-b16f-5f91be1718f7